// src/app/pages/fna-bram/inputs/input-data-dialog/input.data.enum.ts
export enum INBOUND_LOGISTICS {
    WAREHOUSE = "WAREHOUSE",
    RAW_MATERIAL = "RAW_MATERIAL",
    TRANSPORTATION = "TRANSPORTATION",
    OTHERS = "OTHERS"
}

export enum OPERATION {
    FACTORY_BUILDING = 'FACTORY_BUILDING',
    MACHINERY = 'MACHINERY',
    MANUFACTURING = 'MANUFACTURING',
    OTHERS = 'OTHERS'
}

export enum SHIPPING_TYPE {
    INLAND = 'INLAND',
    COASTAL = 'COASTAL'
}

export enum OUTBOUND_LOGISTICS {
    WAREHOUSING = 'WAREHOUSING',
    GOODS_WAREHOUSE = 'GOODS_WAREHOUSE',
    GOODS_TRANSPORTATION = 'GOODS_TRANSPORTATION',
    TRANSPORATION = 'TRANSPORATION',
    OTHERS = 'OTHERS'
}

export enum MARKETING_SALES {
    MARKETING_ASSETS = 'MARKETING_ASSETS',
    SALES_TEAM = 'SALES_TEAM',
    TRANSPORATION = 'TRANSPORATION',
    OTHERS = 'OTHERS'
}

export enum SERVICE {
    SERVICE_INFRASTRUCTURER = 'SERVICE_INFRASTRUCTURER',
    CALL_CENTER = 'CALL_CENTER',
    SERVICING_EMPLOYEES = 'SERVICING_EMPLOYEES',
    OTHERS = 'OTHERS'
}

export enum TECHNOLOGY {
    HARDWARE = 'HARDWARE',
    SYSTEM_SOFTWARE = 'SYSTEM_SOFTWARE',
    CUST_DATA_INFORMATION = 'CUST_DATA_INFORMATION',
    OTHER = 'OTHER'
}

export enum HUMAN_RESOURCES {
    RECRUITMENT_AND_TRAINING = 'RECRUITMENT_AND_TRAINING'
}

export enum MANAGEMENT_INFRASTRUCTURE {
    MANAGEMENT_TEAM = 'MANAGEMENT_TEAM',
    LOCATION_BUSINESS = 'LOCATION_BUSINESS',
    FIX_ASSET_MANAGEMENT = 'FIX_ASSET_MANAGEMENT'
}